+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-# Release TBD
+## HEAD -> main (Release TBD)
 
 - Added
   - Copy Value At Marker context menu item